Extreme optimizations and performance improvements
By initializing a value (in this case 'qt'), you are causing the JavaScript engine to allocate memory for that specific integer which could cause the garbage collector to slow down by `1×(10 to the power of -13)ms`  which in my opinion is totally unacceptable since the method `cute` returns a static value that never changes, and Lumap is always cute so the end user would never attempt to change that value.

Armed with this information, this PR will remove the `qt` property and change the `cute` method to simply return true. Since we now no longer use the `this` keyword, there is no need to add a constructor and call `super` on the parent class (`Map`) which could massively save CPU time on low-end machines.

That said, I have tested everything and have immediately noticed extreme performance boosts and now the package could run on my WW2 toaster when it previously could not.